[MODI-85] REFACTOR:  유저 조회 API 엔드포인트 변경
